+++ facebook/src/base_facebook.php
@@ -37,7 +37,7 @@ class FacebookApiException extends Exception
   /**
    * Make a new API Exception with the given result.
    *
-   * @param Array $result the result from the API server
+   * @param array $result The result from the API server
    */
   public function __construct($result) {
     $this->result = $result;
@@ -63,7 +63,7 @@ class FacebookApiException extends Exception
   /**
    * Return the associated result object returned by the API server.
    *
-   * @returns Array the result from the API server
+   * @return array The result from the API server
    */
   public function getResult() {
     return $this->result;
@@ -73,7 +73,7 @@ class FacebookApiException extends Exception
    * Returns the associated type for the error. This will default to
    * 'Exception' when a type is not available.
    *
-   * @return String
+   * @return string
    */
   public function getType() {
     if (isset($this->result['error'])) {
@@ -95,7 +95,7 @@ class FacebookApiException extends Exception
   /**
    * To make debugging easier.
    *
-   * @returns String the string representation of the error
+   * @return string The string representation of the error
    */
   public function __toString() {
     $str = $this->getType() . ': ';
@@ -109,8 +109,8 @@ class FacebookApiException extends Exception
 /**
  * Provides access to the Facebook Platform.  This class provides
  * a majority of the functionality needed, but the class is abstract
- * because it is designed to be subclassed.  The subclass must
- * implement the three abstract methods listed at the bottom of
+ * because it is designed to be sub-classed.  The subclass must
+ * implement the four abstract methods listed at the bottom of
  * the file.
  *
  * @author Naitik Shah <naitik@facebook.com>
@@ -120,7 +120,12 @@ abstract class BaseFacebook
   /**
    * Version.
    */
-  const VERSION = '3.0.0';
+  const VERSION = '3.2.0';
+
+  /**
+   * Signed Request Algorithm.
+   */
+  const SIGNED_REQUEST_ALGORITHM = 'HMAC-SHA256';
 
   /**
    * Default options for curl.
@@ -129,7 +134,7 @@ abstract class BaseFacebook
     CURLOPT_CONNECTTIMEOUT => 10,
     CURLOPT_RETURNTRANSFER => true,
     CURLOPT_TIMEOUT        => 60,
-    CURLOPT_USERAGENT      => 'facebook-php-3.0',
+    CURLOPT_USERAGENT      => 'facebook-php-3.2',
   );
 
   /**
@@ -146,25 +151,32 @@ abstract class BaseFacebook
    * Maps aliases to Facebook domains.
    */
   public static $DOMAIN_MAP = array(
-    'api'       => 'https://api.facebook.com/',
-    'api_video' => 'https://api-video.facebook.com/',
-    'api_read'  => 'https://api-read.facebook.com/',
-    'graph'     => 'https://graph.facebook.com/',
-    'www'       => 'https://www.facebook.com/',
+    'api'         => 'https://api.facebook.com/',
+    'api_video'   => 'https://api-video.facebook.com/',
+    'api_read'    => 'https://api-read.facebook.com/',
+    'graph'       => 'https://graph.facebook.com/',
+    'graph_video' => 'https://graph-video.facebook.com/',
+    'www'         => 'https://www.facebook.com/',
   );
 
   /**
    * The Application ID.
+   *
+   * @var string
    */
   protected $appId;
 
   /**
-   * The Application API Secret.
+   * The Application App Secret.
+   *
+   * @var string
    */
-  protected $apiSecret;
+  protected $appSecret;
 
   /**
    * The ID of the Facebook user, or 0 if the user is logged out.
+   *
+   * @var integer
    */
   protected $user;
 
@@ -174,21 +186,31 @@ abstract class BaseFacebook
   protected $signedRequest;
 
   /**
-   * A CSRF state variable to assist in the defense against
-   * CSRF attacks.
+   * A CSRF state variable to assist in the defense against CSRF attacks.
    */
   protected $state;
 
   /**
    * The OAuth access token received in exchange for a valid authorization
    * code.  null means the access token has yet to be determined.
+   *
+   * @var string
    */
   protected $accessToken = null;
 
   /**
    * Indicates if the CURL based @ syntax for file uploads is enabled.
+   *
+   * @var boolean
    */
   protected $fileUploadSupport = false;
+
+  /**
+   * Indicates if we trust HTTP_X_FORWARDED_* headers.
+   *
+   * @var boolean
+   */
+  protected $trustForwarded = false;
 
   /**
    * Initialize a Facebook Application.
@@ -198,23 +220,28 @@ abstract class BaseFacebook
    * - secret: the application secret
    * - fileUpload: (optional) boolean indicating if file uploads are enabled
    *
-   * @param Array $config the application configuration
+   * @param array $config The application configuration
    */
   public function __construct($config) {
     $this->setAppId($config['appId']);
-    $this->setApiSecret($config['secret']);
+    $this->setAppSecret($config['secret']);
     if (isset($config['fileUpload'])) {
       $this->setFileUploadSupport($config['fileUpload']);
     }
-    if (isset($_COOKIE[$this->getCSRFTokenCookieName()])) {
-      $this->state = $_COOKIE[$this->getCSRFTokenCookieName()];
+    if (isset($config['trustForwarded']) && $config['trustForwarded']) {
+      $this->trustForwarded = true;
+    }
+    $state = $this->getPersistentData('state');
+    if (!empty($state)) {
+      $this->state = $state;
     }
   }
 
   /**
    * Set the Application ID.
    *
-   * @param String $appId the Application ID
+   * @param string $appId The Application ID
+   * @return BaseFacebook
    */
   public function setAppId($appId) {
     $this->appId = $appId;
@@ -224,35 +251,59 @@ abstract class BaseFacebook
   /**
    * Get the Application ID.
    *
-   * @return String the Application ID
+   * @return string the Application ID
    */
   public function getAppId() {
     return $this->appId;
   }
 
   /**
-   * Set the API Secret.
+   * Set the App Secret.
    *
-   * @param String $appId the API Secret
+   * @param string $apiSecret The App Secret
+   * @return BaseFacebook
+   * @deprecated
    */
   public function setApiSecret($apiSecret) {
-    $this->apiSecret = $apiSecret;
+    $this->setAppSecret($apiSecret);
     return $this;
   }
 
   /**
-   * Get the API Secret.
+   * Set the App Secret.
    *
-   * @return String the API Secret
+   * @param string $appSecret The App Secret
+   * @return BaseFacebook
+   */
+  public function setAppSecret($appSecret) {
+    $this->appSecret = $appSecret;
+    return $this;
+  }
+
+  /**
+   * Get the App Secret.
+   *
+   * @return string the App Secret
+   * @deprecated
    */
   public function getApiSecret() {
-    return $this->apiSecret;
+    return $this->getAppSecret();
+  }
+
+  /**
+   * Get the App Secret.
+   *
+   * @return string the App Secret
+   */
+  public function getAppSecret() {
+    return $this->appSecret;
   }
 
   /**
    * Set the file upload support status.
    *
-   * @param Boolean the file upload support status.
+   * @param boolean $fileUploadSupport The file upload support status.
+   * @return BaseFacebook
    */
   public function setFileUploadSupport($fileUploadSupport) {
     $this->fileUploadSupport = $fileUploadSupport;
@@ -262,11 +313,21 @@ abstract class BaseFacebook
   /**
    * Get the file upload support status.
    *
-   * @return Boolean true if and only if the server supports
-   * file upload.
+   * @return boolean true if and only if the server supports file upload.
+   */
+  public function getFileUploadSupport() {
+    return $this->fileUploadSupport;
+  }
+
+  /**
+   * DEPRECATED! Please use getFileUploadSupport instead.
+   *
+   * Get the file upload support status.
+   *
+   * @return boolean true if and only if the server supports file upload.
    */
   public function useFileUploadSupport() {
-    return $this->fileUploadSupport;
+    return $this->getFileUploadSupport();
   }
 
   /**
@@ -274,11 +335,55 @@ abstract class BaseFacebook
    * your access token by other means and just want the SDK
    * to use it.
    *
-   * @param String $access_token an access token.
+   * @param string $access_token an access token.
+   * @return BaseFacebook
    */
   public function setAccessToken($access_token) {
     $this->accessToken = $access_token;
     return $this;
+  }
+
+  /**
+   * Extend an access token, while removing the short-lived token that might
+   * have been generated via client-side flow. Thanks to http://bit.ly/b0Pt0H
+   * for the workaround.
+   */
+  public function setExtendedAccessToken() {
+    try {
+      // need to circumvent json_decode by calling _oauthRequest
+      // directly, since response isn't JSON format.
+      $access_token_response = $this->_oauthRequest(
+        $this->getUrl('graph', '/oauth/access_token'),
+        $params = array(
+          'client_id' => $this->getAppId(),
+          'client_secret' => $this->getAppSecret(),
+          'grant_type' => 'fb_exchange_token',
+          'fb_exchange_token' => $this->getAccessToken(),
+        )
+      );
+    }
+    catch (FacebookApiException $e) {
+      // most likely that user very recently revoked authorization.
+      // In any event, we don't have an access token, so say so.
+      return false;
+    }
+  
+    if (empty($access_token_response)) {
+      return false;
+    }
+      
+    $response_params = array();
+    parse_str($access_token_response, $response_params);
+    
+    if (!isset($response_params['access_token'])) {
+      return false;
+    }
+    
+    $this->destroySession();
+    
+    $this->setPersistentData(
+      'access_token', $response_params['access_token']
+    );
   }
 
   /**
@@ -288,7 +393,7 @@ abstract class BaseFacebook
    * access token if a valid user access token wasn't available.  Subsequent
    * calls return whatever the first call returned.
    *
-   * @return String the access token
+   * @return string The access token
    */
   public function getAccessToken() {
     if ($this->accessToken !== null) {
@@ -300,7 +405,8 @@ abstract class BaseFacebook
     // access token, in case we navigate to the /oauth/access_token
     // endpoint, where SOME access token is required.
     $this->setAccessToken($this->getApplicationAccessToken());
-    if ($user_access_token = $this->getUserAccessToken()) {
+    $user_access_token = $this->getUserAccessToken();
+    if ($user_access_token) {
       $this->setAccessToken($user_access_token);
     }
 
@@ -314,8 +420,8 @@ abstract class BaseFacebook
    * return a valid user access token, or false if one is determined
    * to not be available.
    *
-   * @return String a valid user access token, or false if one
-   *         could not be determined.
+   * @return string A valid user access token, or false if one
+   *                could not be determined.
    */
   protected function getUserAccessToken() {
     // first, consider a signed request if it's supplied.
@@ -323,10 +429,22 @@ abstract class BaseFacebook
     // the access token.
     $signed_request = $this->getSignedRequest();
     if ($signed_request) {
+      // apps.facebook.com hands the access_token in the signed_request
       if (array_key_exists('oauth_token', $signed_request)) {
         $access_token = $signed_request['oauth_token'];
         $this->setPersistentData('access_token', $access_token);
         return $access_token;
+      }
+
+      // the JS SDK puts a code in with the redirect_uri of ''
+      if (array_key_exists('code', $signed_request)) {
+        $code = $signed_request['code'];
+        $access_token = $this->getAccessTokenFromCode($code, '');
+        if ($access_token) {
+          $this->setPersistentData('code', $code);
+          $this->setPersistentData('access_token', $access_token);
+          return $access_token;
+        }
       }
 
       // signed request states there's no access token, so anything
@@ -358,15 +476,19 @@ abstract class BaseFacebook
   }
 
   /**
-   * Get the data from a signed_request token.
+   * Retrieve the signed request, either from a request parameter or,
+   * if not present, from a cookie.
    *
-   * @return String the base domain
+   * @return string the signed request, if available, or null otherwise.
    */
   public function getSignedRequest() {
     if (!$this->signedRequest) {
       if (isset($_REQUEST['signed_request'])) {
         $this->signedRequest = $this->parseSignedRequest(
           $_REQUEST['signed_request']);
+      } else if (isset($_COOKIE[$this->getSignedRequestCookieName()])) {
+        $this->signedRequest = $this->parseSignedRequest(
+          $_COOKIE[$this->getSignedRequestCookieName()]);
       }
     }
     return $this->signedRequest;
@@ -376,7 +498,7 @@ abstract class BaseFacebook
    * Get the UID of the connected user, or 0
    * if the Facebook user is not connected.
    *
-   * @return String the UID if available.
+   * @return string the UID if available.
    */
   public function getUser() {
     if ($this->user !== null) {
@@ -392,8 +514,8 @@ abstract class BaseFacebook
    * requests, then considering an authorization code, and then
    * falling back to any persistent store storing the user.
    *
-   * @return Integer the id of the connected Facebook user, or 0
-   * if no such user exists.
+   * @return integer The id of the connected Facebook user,
+   *                 or 0 if no such user exists.
    */
   protected function getUserFromAvailableData() {
     // if a signed request is supplied, then it solely determines
@@ -441,12 +563,19 @@ abstract class BaseFacebook
    * - redirect_uri: the url to go to after a successful login
    * - scope: comma separated list of requested extended perms
    *
-   * @param Array $params provide custom parameters
-   * @return String the URL for the login flow
+   * @param array $params Provide custom parameters
+   * @return string The URL for the login flow
    */
   public function getLoginUrl($params=array()) {
     $this->establishCSRFTokenState();
     $currentUrl = $this->getCurrentUrl();
+
+    // if 'scope' is passed as an array, convert to comma separated list
+    $scopeParams = isset($params['scope']) ? $params['scope'] : null;
+    if ($scopeParams && is_array($scopeParams)) {
+      $params['scope'] = implode(',', $scopeParams);
+    }
+
     return $this->getUrl(
       'www',
       'dialog/oauth',
@@ -463,8 +592,8 @@ abstract class BaseFacebook
    * The parameters:
    * - next: the url to go to after a successful logout
    *
-   * @param Array $params provide custom parameters
-   * @return String the URL for the logout flow
+   * @param array $params Provide custom parameters
+   * @return string The URL for the logout flow
    */
   public function getLogoutUrl($params=array()) {
     return $this->getUrl(
@@ -472,7 +601,7 @@ abstract class BaseFacebook
       'logout.php',
       array_merge(array(
         'next' => $this->getCurrentUrl(),
-        'access_token' => $this->getAccessToken(),
+        'access_token' => $this->getUserAccessToken(),
       ), $params)
     );
   }
@@ -485,8 +614,8 @@ abstract class BaseFacebook
    * - no_session: the URL to go to if the user is not connected
    * - no_user: the URL to go to if the user is not signed into facebook
    *
-   * @param Array $params provide custom parameters
-   * @return String the URL for the logout flow
+   * @param array $params Provide custom parameters
+   * @return string The URL for the logout flow
    */
   public function getLoginStatusUrl($params=array()) {
     return $this->getUrl(
@@ -505,8 +634,7 @@ abstract class BaseFacebook
   /**
    * Make an API call.
    *
-   * @param Array $params the API call parameters
-   * @return the decoded response
+   * @return mixed The decoded response
    */
   public function api(/* polymorphic */) {
     $args = func_get_args();
@@ -518,12 +646,36 @@ abstract class BaseFacebook
   }
 
   /**
+   * Constructs and returns the name of the cookie that
+   * potentially houses the signed request for the app user.
+   * The cookie is not set by the BaseFacebook class, but
+   * it may be set by the JavaScript SDK.
+   *
+   * @return string the name of the cookie that would house
+   *         the signed request value.
+   */
+  protected function getSignedRequestCookieName() {
+    return 'fbsr_'.$this->getAppId();
+  }
+
+  /**
+   * Constructs and returns the name of the coookie that potentially contain
+   * metadata. The cookie is not set by the BaseFacebook class, but it may be
+   * set by the JavaScript SDK.
+   *
+   * @return string the name of the cookie that would house metadata.
+   */
+  protected function getMetadataCookieName() {
+    return 'fbm_'.$this->getAppId();
+  }
+
+  /**
    * Get the authorization code from the query parameters, if it exists,
    * and otherwise return false to signal no authorization code was
    * discoverable.
    *
-   * @return mixed the authorization code, or false if the authorization
-   * code could not be determined.
+   * @return mixed The authorization code, or false if the authorization
+   *               code could not be determined.
    */
   protected function getCode() {
     if (isset($_REQUEST['code'])) {
@@ -533,7 +685,7 @@ abstract class BaseFacebook
 
         // CSRF state has done its job, so clear it
         $this->state = null;
-        unset($_COOKIE[$this->getCSRFTokenCookieName()]);
+        $this->clearPersistentData('state');
         return $_REQUEST['code'];
       } else {
         self::errorLog('CSRF state token does not match one provided.');
@@ -548,11 +700,11 @@ abstract class BaseFacebook
    * Retrieves the UID with the understanding that
    * $this->accessToken has already been set and is
    * seemingly legitimate.  It relies on Facebook's Graph API
-   * to retreive user information and then extract
+   * to retrieve user information and then extract
    * the user ID.
    *
-   * @return Integer returns the UID of the Facebook user, or
-   * 0 if the Facebook user could not be determined.
+   * @return integer Returns the UID of the Facebook user, or 0
+   *                 if the Facebook user could not be determined.
    */
   protected function getUserFromAccessToken() {
     try {
@@ -567,42 +719,44 @@ abstract class BaseFacebook
    * Returns the access token that should be used for logged out
    * users when no authorization code is available.
    *
-   * @return String the application access token, useful for
-   * gathering public information about users and applications.
+   * @return string The application access token, useful for gathering
+   *                public information about users and applications.
    */
   protected function getApplicationAccessToken() {
-    return $this->appId.'|'.$this->apiSecret;
+    return $this->appId.'|'.$this->appSecret;
   }
 
   /**
-   * Lays down a CSRF state token for this process.  We
-   * only generate a new CSRF token if one isn't currently
-   * circulating in the domain's cookie jar.
+   * Lays down a CSRF state token for this process.
    *
    * @return void
    */
   protected function establishCSRFTokenState() {
     if ($this->state === null) {
       $this->state = md5(uniqid(mt_rand(), true));
-      @setcookie($this->getCSRFTokenCookieName(),$this->state,(time() + 3600)); // sticks for an hour
+      $this->setPersistentData('state', $this->state);
     }
   }
 
   /**
-   * Retreives an access token for the given authorization code
+   * Retrieves an access token for the given authorization code
    * (previously generated from www.facebook.com on behalf of
    * a specific user).  The authorization code is sent to graph.facebook.com
    * and a legitimate access token is generated provided the access token
    * and the user for which it was generated all match, and the user is
    * either logged in to Facebook or has granted an offline access permission.
    *
-   * @param String $code an authorization code.
-   * @return mixed an access token exchanged for the authorization code, or
-   * false if an access token could not be generated.
+   * @param string $code An authorization code.
+   * @return mixed An access token exchanged for the authorization code, or
+   *               false if an access token could not be generated.
    */
-  protected function getAccessTokenFromCode($code) {
+  protected function getAccessTokenFromCode($code, $redirect_uri = null) {
     if (empty($code)) {
       return false;
+    }
+
+    if ($redirect_uri === null) {
+      $redirect_uri = $this->getCurrentUrl();
     }
 
     try {
@@ -612,8 +766,8 @@ abstract class BaseFacebook
         $this->_oauthRequest(
           $this->getUrl('graph', '/oauth/access_token'),
           $params = array('client_id' => $this->getAppId(),
-                          'client_secret' => $this->getApiSecret(),
-                          'redirect_uri' => $this->getCurrentUrl(),
+                          'client_secret' => $this->getAppSecret(),
+                          'redirect_uri' => $redirect_uri,
                           'code' => $code));
     } catch (FacebookApiException $e) {
       // most likely that user very recently revoked authorization.
@@ -637,8 +791,9 @@ abstract class BaseFacebook
   /**
    * Invoke the old restserver.php endpoint.
    *
-   * @param Array $params method call object
-   * @return the decoded response object
+   * @param array $params Method call object
+   *
+   * @return mixed The decoded response object
    * @throws FacebookApiException
    */
   protected function _restserver($params) {
@@ -653,19 +808,43 @@ abstract class BaseFacebook
 
     // results are returned, errors are thrown
     if (is_array($result) && isset($result['error_code'])) {
-      throw new FacebookApiException($result);
+      $this->throwAPIException($result);
+      // @codeCoverageIgnoreStart
+    }
+    // @codeCoverageIgnoreEnd
+
+    $method = strtolower($params['method']);
+    if ($method === 'auth.expiresession' ||
+        $method === 'auth.revokeauthorization') {
+      $this->destroySession();
     }
 
     return $result;
   }
 
   /**
+   * Return true if this is video post.
+   *
+   * @param string $path The path
+   * @param string $method The http method (default 'GET')
+   *
+   * @return boolean true if this is video post
+   */
+  protected function isVideoPost($path, $method = 'GET') {
+    if ($method == 'POST' && preg_match("/^(\/)(.+)(\/)(videos)$/", $path)) {
+      return true;
+    }
+    return false;
+  }
+
+  /**
    * Invoke the Graph API.
    *
-   * @param String $path the path (required)
-   * @param String $method the http method (default 'GET')
-   * @param Array $params the query/post data
-   * @return the decoded response object
+   * @param string $path The path (required)
+   * @param string $method The http method (default 'GET')
+   * @param array $params The query/post data
+   *
+   * @return mixed The decoded response object
    * @throws FacebookApiException
    */
   protected function _graph($path, $method = 'GET', $params = array()) {
@@ -675,15 +854,23 @@ abstract class BaseFacebook
     }
     $params['method'] = $method; // method override as we always do a POST
 
+    if ($this->isVideoPost($path, $method)) {
+      $domainKey = 'graph_video';
+    } else {
+      $domainKey = 'graph';
+    }
+
     $result = json_decode($this->_oauthRequest(
-      $this->getUrl('graph', $path),
+      $this->getUrl($domainKey, $path),
       $params
     ), true);
 
     // results are returned, errors are thrown
     if (is_array($result) && isset($result['error'])) {
       $this->throwAPIException($result);
+      // @codeCoverageIgnoreStart
     }
+    // @codeCoverageIgnoreEnd
 
     return $result;
   }
@@ -691,9 +878,10 @@ abstract class BaseFacebook
   /**
    * Make a OAuth Request.
    *
-   * @param String $path the path (required)
-   * @param Array $params the query/post data
-   * @return the decoded response object
+   * @param string $url The path (required)
+   * @param array $params The query/post data
+   *
+   * @return string The decoded response object
    * @throws FacebookApiException
    */
   protected function _oauthRequest($url, $params) {
@@ -712,14 +900,15 @@ abstract class BaseFacebook
   }
 
   /**
-   * Makes an HTTP request. This method can be overriden by subclasses if
+   * Makes an HTTP request. This method can be overridden by subclasses if
    * developers want to do fancier things or use something other than curl to
    * make the request.
    *
-   * @param String $url the URL to make the request to
-   * @param Array $params the parameters to use for the POST body
-   * @param CurlHandler $ch optional initialized curl handle
-   * @return String the response text
+   * @param string $url The URL to make the request to
+   * @param array $params The parameters to use for the POST body
+   * @param CurlHandler $ch Initialized curl handle
+   *
+   * @return string The response text
    */
   protected function makeRequest($url, $params, $ch=null) {
     if (!$ch) {
@@ -727,7 +916,7 @@ abstract class BaseFacebook
     }
 
     $opts = self::$CURL_OPTS;
-    if ($this->useFileUploadSupport()) {
+    if ($this->getFileUploadSupport()) {
       $opts[CURLOPT_POSTFIELDS] = $params;
     } else {
       $opts[CURLOPT_POSTFIELDS] = http_build_query($params, null, '&');
@@ -755,6 +944,25 @@ abstract class BaseFacebook
       $result = curl_exec($ch);
     }
 
+    // With dual stacked DNS responses, it's possible for a server to
+    // have IPv6 enabled but not have IPv6 connectivity.  If this is
+    // the case, curl will try IPv4 first and if that fails, then it will
+    // fall back to IPv6 and the error EHOSTUNREACH is returned by the
+    // operating system.
+    if ($result === false && empty($opts[CURLOPT_IPRESOLVE])) {
+        $matches = array();
+        $regex = '/Failed to connect to ([^:].*): Network is unreachable/';
+        if (preg_match($regex, curl_error($ch), $matches)) {
+          if (strlen(@inet_pton($matches[1])) === 16) {
+            self::errorLog('Invalid IPv6 configuration on server, '.
+                           'Please disable or get native IPv6 on your server.');
+            self::$CURL_OPTS[CURLOPT_IPRESOLVE] = CURL_IPRESOLVE_V4;
+            curl_setopt($ch, CURLOPT_IPRESOLVE, CURL_IPRESOLVE_V4);
+            $result = curl_exec($ch);
+          }
+        }
+    }
+
     if ($result === false) {
       $e = new FacebookApiException(array(
         'error_code' => curl_errno($ch),
@@ -771,19 +979,10 @@ abstract class BaseFacebook
   }
 
   /**
-   * The name of the cookie housing the CSRF protection value.
-   *
-   * @return String the cookie name
-   */
-  protected function getCSRFTokenCookieName() {
-    return 'fbcsrf_'.$this->getAppId();
-  }
-
-  /**
    * Parses a signed_request and validates the signature.
    *
-   * @param String A signed token
-   * @return Array the payload inside it or null if the sig is wrong
+   * @param string $signed_request A signed token
+   * @return array The payload inside it or null if the sig is wrong
    */
   protected function parseSignedRequest($signed_request) {
     list($encoded_sig, $payload) = explode('.', $signed_request, 2);
@@ -792,14 +991,15 @@ abstract class BaseFacebook
     $sig = self::base64UrlDecode($encoded_sig);
     $data = json_decode(self::base64UrlDecode($payload), true);
 
-    if (strtoupper($data['algorithm']) !== 'HMAC-SHA256') {
-      self::errorLog('Unknown algorithm. Expected HMAC-SHA256');
+    if (strtoupper($data['algorithm']) !== self::SIGNED_REQUEST_ALGORITHM) {
+      self::errorLog(
+        'Unknown algorithm. Expected ' . self::SIGNED_REQUEST_ALGORITHM);
       return null;
     }
 
     // check sig
     $expected_sig = hash_hmac('sha256', $payload,
-                              $this->getApiSecret(), $raw = true);
+                              $this->getAppSecret(), $raw = true);
     if ($sig !== $expected_sig) {
       self::errorLog('Bad Signed JSON signature!');
       return null;
@@ -809,10 +1009,30 @@ abstract class BaseFacebook
   }
 
   /**
+   * Makes a signed_request blob using the given data.
+   *
+   * @param array The data array.
+   * @return string The signed request.
+   */
+  protected function makeSignedRequest($data) {
+    if (!is_array($data)) {
+      throw new InvalidArgumentException(
+        'makeSignedRequest expects an array. Got: ' . print_r($data, true));
+    }
+    $data['algorithm'] = self::SIGNED_REQUEST_ALGORITHM;
+    $data['issued_at'] = time();
+    $json = json_encode($data);
+    $b64 = self::base64UrlEncode($json);
+    $raw_sig = hash_hmac('sha256', $b64, $this->getAppSecret(), $raw = true);
+    $sig = self::base64UrlEncode($raw_sig);
+    return $sig.'.'.$b64;
+  }
+
+  /**
    * Build the URL for api given parameters.
    *
    * @param $method String the method name.
-   * @return String the URL for the given parameters
+   * @return string The URL for the given parameters
    */
   protected function getApiUrl($method) {
     static $READ_ONLY_CALLS =
@@ -888,10 +1108,11 @@ abstract class BaseFacebook
   /**
    * Build the URL for given domain alias, path and parameters.
    *
-   * @param $name String the name of the domain
-   * @param $path String optional path (without a leading slash)
-   * @param $params Array optional query parameters
-   * @return String the URL for the given parameters
+   * @param $name string The name of the domain
+   * @param $path string Optional path (without a leading slash)
+   * @param $params array Optional query parameters
+   *
+   * @return string The URL for the given parameters
    */
   protected function getUrl($name, $path='', $params=array()) {
     $url = self::$DOMAIN_MAP[$name];
@@ -908,29 +1129,68 @@ abstract class BaseFacebook
     return $url;
   }
 
+  protected function getHttpHost() {
+    if ($this->trustForwarded && isset($_SERVER['HTTP_X_FORWARDED_HOST'])) {
+      return $_SERVER['HTTP_X_FORWARDED_HOST'];
+    }
+    return $_SERVER['HTTP_HOST'];
+  }
+
+  protected function getHttpProtocol() {
+    if ($this->trustForwarded && isset($_SERVER['HTTP_X_FORWARDED_PROTO'])) {
+      if ($_SERVER['HTTP_X_FORWARDED_PROTO'] === 'https') {
+        return 'https';
+      }
+      return 'http';
+    }
+    if (isset($_SERVER['HTTPS']) &&
+        ($_SERVER['HTTPS'] === 'on' || $_SERVER['HTTPS'] == 1)) {
+      return 'https';
+    }
+    return 'http';
+  }
+
+  /**
+   * Get the base domain used for the cookie.
+   */
+  protected function getBaseDomain() {
+    // The base domain is stored in the metadata cookie if not we fallback
+    // to the current hostname
+    $metadata = $this->getMetadataCookie();
+    if (array_key_exists('base_domain', $metadata) &&
+        !empty($metadata['base_domain'])) {
+      return trim($metadata['base_domain'], '.');
+    }
+    return $this->getHttpHost();
+  }
+
+  /**
+
   /**
    * Returns the Current URL, stripping it of known FB parameters that should
    * not persist.
    *
-   * @return String the current URL
+   * @return string The current URL
    */
   protected function getCurrentUrl() {
-    $protocol = isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] == 'on'
-      ? 'https://'
-      : 'http://';
-    $currentUrl = $protocol . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
+    $protocol = $this->getHttpProtocol() . '://';
+    $host = $this->getHttpHost();
+    $currentUrl = $protocol.$host.$_SERVER['REQUEST_URI'];
     $parts = parse_url($currentUrl);
 
-    // drop known fb params
     $query = '';
     if (!empty($parts['query'])) {
-      $params = array();
-      parse_str($parts['query'], $params);
-      foreach(self::$DROP_QUERY_PARAMS as $key) {
-        unset($params[$key]);
+      // drop known fb params
+      $params = explode('&', $parts['query']);
+      $retained_params = array();
+      foreach ($params as $param) {
+        if ($this->shouldRetainParam($param)) {
+          $retained_params[] = $param;
+        }
       }
-      if (!empty($params)) {
-        $query = '?' . http_build_query($params, null, '&');
+
+      if (!empty($retained_params)) {
+        $query = '?'.implode($retained_params, '&');
       }
     }
 
@@ -946,12 +1206,33 @@ abstract class BaseFacebook
   }
 
   /**
+   * Returns true if and only if the key or key/value pair should
+   * be retained as part of the query string.  This amounts to
+   * a brute-force search of the very small list of Facebook-specific
+   * params that should be stripped out.
+   *
+   * @param string $param A key or key/value pair within a URL's query (e.g.
+   *                     'foo=a', 'foo=', or 'foo'.
+   *
+   * @return boolean
+   */
+  protected function shouldRetainParam($param) {
+    foreach (self::$DROP_QUERY_PARAMS as $drop_query_param) {
+      if (strpos($param, $drop_query_param.'=') === 0) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  /**
    * Analyzes the supplied result to see if it was thrown
    * because the access token is no longer valid.  If that is
-   * the case, then the persistent store is cleared.
+   * the case, then we destroy the session.
    *
-   * @param $result a record storing the error message returned
-   *        by a failed API call.
+   * @param $result array A record storing the error message returned
+   *                      by a failed API call.
    */
   protected function throwAPIException($result) {
     $e = new FacebookApiException($result);
@@ -960,13 +1241,16 @@ abstract class BaseFacebook
       case 'OAuthException':
         // OAuth 2.0 Draft 10 style
       case 'invalid_token':
+        // REST server errors are just Exceptions
+      case 'Exception':
         $message = $e->getMessage();
-      if ((strpos($message, 'Error validating access token') !== false) ||
-          (strpos($message, 'Invalid OAuth access token') !== false)) {
-        $this->setAccessToken(null);
-        $this->user = 0;
-        $this->clearAllPersistentData();
-      }
+        if ((strpos($message, 'Error validating access token') !== false) ||
+            (strpos($message, 'Invalid OAuth access token') !== false) ||
+            (strpos($message, 'An active access token must be used') !== false)
+        ) {
+          $this->destroySession();
+        }
+        break;
     }
 
     throw $e;
@@ -976,7 +1260,7 @@ abstract class BaseFacebook
   /**
    * Prints to the error log if you aren't in command line mode.
    *
-   * @param String log message
+   * @param string $msg Log message
    */
   protected static function errorLog($msg) {
     // disable error log if we are running in a CLI environment
@@ -994,15 +1278,107 @@ abstract class BaseFacebook
    * Exactly the same as base64_encode except it uses
    *   - instead of +
    *   _ instead of /
+   *   No padded =
    *
-   * @param String base64UrlEncodeded string
+   * @param string $input base64UrlEncoded string
+   * @return string
    */
   protected static function base64UrlDecode($input) {
     return base64_decode(strtr($input, '-_', '+/'));
   }
 
   /**
-   * Each of the following three methods should be overridden in
+   * Base64 encoding that doesn't need to be urlencode()ed.
+   * Exactly the same as base64_encode except it uses
+   *   - instead of +
+   *   _ instead of /
+   *
+   * @param string $input string
+   * @return string base64Url encoded string
+   */
+  protected static function base64UrlEncode($input) {
+    $str = strtr(base64_encode($input), '+/', '-_');
+    $str = str_replace('=', '', $str);
+    return $str;
+  }
+
+  /**
+   * Destroy the current session
+   */
+  public function destroySession() {
+    $this->accessToken = null;
+    $this->signedRequest = null;
+    $this->user = null;
+    $this->clearAllPersistentData();
+
+    // Javascript sets a cookie that will be used in getSignedRequest that we
+    // need to clear if we can
+    $cookie_name = $this->getSignedRequestCookieName();
+    if (array_key_exists($cookie_name, $_COOKIE)) {
+      unset($_COOKIE[$cookie_name]);
+      if (!headers_sent()) {
+        $base_domain = $this->getBaseDomain();
+        setcookie($cookie_name, '', 1, '/', '.'.$base_domain);
+      } else {
+        // @codeCoverageIgnoreStart
+        self::errorLog(
+          'There exists a cookie that we wanted to clear that we couldn\'t '.
+          'clear because headers was already sent. Make sure to do the first '.
+          'API call before outputing anything.'
+        );
+        // @codeCoverageIgnoreEnd
+      }
+    }
+  }
+
+  /**
+   * Parses the metadata cookie that our Javascript API set
+   *
+   * @return  an array mapping key to value
+   */
+  protected function getMetadataCookie() {
+    $cookie_name = $this->getMetadataCookieName();
+    if (!array_key_exists($cookie_name, $_COOKIE)) {
+      return array();
+    }
+
+    // The cookie value can be wrapped in "-characters so remove them
+    $cookie_value = trim($_COOKIE[$cookie_name], '"');
+
+    if (empty($cookie_value)) {
+      return array();
+    }
+
+    $parts = explode('&', $cookie_value);
+    $metadata = array();
+    foreach ($parts as $part) {
+      $pair = explode('=', $part, 2);
+      if (!empty($pair[0])) {
+        $metadata[urldecode($pair[0])] =
+          (count($pair) > 1) ? urldecode($pair[1]) : '';
+      }
+    }
+
+    return $metadata;
+  }
+
+  protected static function isAllowedDomain($big, $small) {
+    if ($big === $small) {
+      return true;
+    }
+    return self::endsWith($big, '.'.$small);
+  }
+
+  protected static function endsWith($big, $small) {
+    $len = strlen($small);
+    if ($len === 0) {
+      return true;
+    }
+    return substr($big, -$len) === $small;
+  }
+
+  /**
+   * Each of the following four methods should be overridden in
    * a concrete subclass, as they are in the provided Facebook class.
    * The Facebook class uses PHP sessions to provide a primitive
    * persistent store, but another subclass--one that you implement--
@@ -1010,7 +1386,40 @@ abstract class BaseFacebook
    *
    * @see Facebook
    */
+
+  /**
+   * Stores the given ($key, $value) pair, so that future calls to
+   * getPersistentData($key) return $value. This call may be in another request.
+   *
+   * @param string $key
+   * @param array $value
+   *
+   * @return void
+   */
   abstract protected function setPersistentData($key, $value);
+
+  /**
+   * Get the data for $key, persisted by BaseFacebook::setPersistentData()
+   *
+   * @param string $key The key of the data to retrieve
+   * @param boolean $default The default value to return if $key is not found
+   *
+   * @return mixed
+   */
   abstract protected function getPersistentData($key, $default = false);
+
+  /**
+   * Clear the data with $key from the persistent storage
+   *
+   * @param string $key
+   * @return void
+   */
+  abstract protected function clearPersistentData($key);
+
+  /**
+   * Clear all data from the persistent storage
+   *
+   * @return void
+   */
   abstract protected function clearAllPersistentData();
 }